+++ src/main/kotlin/com/example/smartshopping/domain/service/SigninService.kt
@@ -1,20 +1,21 @@
-package com.example.smartshopping.domain.auth.signin
+package com.example.smartshopping.domain.service
 
 import com.example.smartshopping.common.SmartShoppingException
 import com.example.smartshopping.domain.auth.JWTUtil
-import com.example.smartshopping.domain.user.User
-import com.example.smartshopping.domain.user.UserRepository
+import com.example.smartshopping.domain.repository.UserRepository
+import com.example.smartshopping.domain.request.SigninRequest
+import com.example.smartshopping.domain.response.SigninResponse
+import com.example.smartshopping.entity.User
 import org.mindrot.jbcrypt.BCrypt
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
-import java.lang.IllegalStateException
 
 @Service
 class SigninService @Autowired constructor(
     private val userRepository: UserRepository
 ) {
 
-    fun signin(signinRequest: SigninRequest) : SigninResponse{
+    fun signin(signinRequest: SigninRequest) : SigninResponse {
         val user = userRepository
             .findByUserId(signinRequest.userId.lowercase())
             ?:throw SmartShoppingException("로그인 정보를 확인해주세요.")
