+++ src/main/kotlin/com/example/smartshopping/domain/service/SignupService.kt
@@ -1,8 +1,9 @@
-package com.example.smartshopping.domain.auth.signup
+package com.example.smartshopping.domain.service
 
 import com.example.smartshopping.common.SmartShoppingException
-import com.example.smartshopping.domain.user.User
-import com.example.smartshopping.domain.user.UserRepository
+import com.example.smartshopping.domain.repository.UserRepository
+import com.example.smartshopping.domain.request.SignupRequest
+import com.example.smartshopping.entity.User
 import org.mindrot.jbcrypt.BCrypt
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
